web/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
python-socketio[client]==5.11.1
requests==2.31.0
python-dotenv==1.0.1
websocket-client==1.7.0